+++ src/main/java/com/baremind/data/WechatUser.java
@@ -41,7 +41,7 @@ public class WechatUser {
     private String nickname;
 
     @Column(name = "sex")
-    private Integer sex;
+    private Long sex;
 
     @Column(name = "city")
     private String city;
@@ -75,6 +75,7 @@ public class WechatUser {
 
     @Column(name = "group_id")
     private int groupId;
+
 
     public String getLanguage() {
         return language;
@@ -188,11 +189,11 @@ public class WechatUser {
         this.nickname = nickname;
     }
 
-    public Integer getSex() {
+    public Long getSex() {
         return sex;
     }
 
-    public void setSex(Integer sex) {
+    public void setSex(Long sex) {
         this.sex = sex;
     }
 
@@ -227,6 +228,14 @@ public class WechatUser {
     public void setHead(String head) {
         this.head = head;
     }
+
+//    public String[] getPrivilege() {
+//        return privilege;
+//    }
+//
+//    public void setPrivilege(String[] privilege) {
+//        this.privilege = privilege;
+//    }
 
     public String getInfo() {
         return info;
